Rename nightly repo: redhat.repo -> rhel8internal.repo
because we have override logic that makes it possible to test an
arbitrary compose. While installing gitlab-ci-runner may be pulling
dependencies from the latest nightly we'd like for the remaining
duration of the test job and installation of osbuild-composer the
repositories to point to the chosen distro under test and make sure
we can override the values if necessary.

See
https://github.com/osbuild/osbuild-composer/blob/main/schutzbot/prepare-rhel-internal.sh