unify main and rhos-01 branches
This commits brings most openstack mages from the rhos-01 branch and puts
them in a new directory called rhos-01. This should allow us to use the same
commit in both rhos-d and rhos-01 repositories. Note that it's still
impossible to mix two RHOS clouds in one repository.

The same pruning of outdated images was done for rhos-01 as in the
previous commit.