aws: convert to spot fleets
Using spot instances directly has one flaw: A specific availability zone might
be out of a specific instance type. In this case, the instance creation fails
and the developer is sad because they need to respin the test.

This commit switches our deployment to use spot fleets. With spot fleets, we
can use multiple zones and instance types and AWS will always pick the
cheapest available option. This is perfect for us as we can run on a wide
range of instance types and AZs.