base: 4h timeout on the spot fleet request
We have a lot of leftovers after CI has run. We need to find a way to
automatically remove outdated EC2 fleet instances. Valid_until is a
variable that if set will timeout the request. With the combination of
"terminate_instances_with_expiration" it should correctly kill the
instance when the 4h mark is reached.