Workaround flaky internal DNS
The DNS resolvers provided by DHCP in AWS are flaky when being queried
for the VSphere instance IP address. This makes the VSphere tests in
osbuild-composer fail a lot on not being able to resolve the hostname.

Add the IP address and hostname to /etc/hosts as a temporary workaround,
until IT fixes the issue.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>